:art: Add the data structure for ipfs config
This is because the python version uses an older version of the api.
The error message using the latest server is as follows

 Python clients require an earlier version of IPFS
  An exception occurred while persisting tweet in IPFS
  Unsupported daemon version '0.12.0' (not in range: 0.4.23 ≤ … < 0.8.0)

image: ipfs/go-ipfs:v0.7.0

In order to improve this, we can just use the property holder to simplify
and document the problem better

For this reason, we run a server that uses the version mentioned.